Move PublicKey to a separate crate
