+++ .gitignore
@@ -10,3 +10,4 @@ wrappers/ios_legacy/vcx/vcx.framework/**
 wrappers/ios_legacy/vcx/vcx.framework.dSYM/**
 .vscode
 **/tails.txt
+.session.vim
