+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
     "did_resolver_registry",
     "did_resolver_sov",
     "did_resolver_web",
+    "public_key",
     "indy_ledger_response_parser",
     "wallet_migrator",
     "tools/simple_message_relay"
