+++ did_peer/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod error;
-pub mod key;
 mod numalgos;
 pub mod peer_did;
 pub mod peer_did_resolver;
