+++ did_peer/src/numalgos/numalgo2/mod.rs
@@ -2,6 +2,7 @@ mod generate;
 mod purpose;
 mod resolve;
 mod service_abbreviated;
+mod verification_method;
 
 pub use generate::generate_numalgo2;
 pub use resolve::resolve_numalgo2;
