+++ public_key/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "public_key"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+thiserror = "1.0.40"
+serde = { version = "1.0.164", features = ["derive"] }
+serde_json = "1.0.96"
+base64 = "0.21.2"
+bs58 = "0.5.0"
+multibase = "0.9.1"
+unsigned-varint = "0.7.1"
