+++ public_key/src/key_type.rs
@@ -1,4 +1,4 @@
-use crate::error::DidPeerError;
+use crate::error::PublicKeyError;
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum KeyType {
@@ -40,7 +40,7 @@ impl From<&KeyType> for u64 {
 }
 
 impl TryFrom<u64> for KeyType {
-    type Error = DidPeerError;
+    type Error = PublicKeyError;
 
     fn try_from(value: u64) -> Result<Self, Self::Error> {
         match value {
@@ -52,7 +52,7 @@ impl TryFrom<u64> for KeyType {
             KeyType::C_P256 => Ok(KeyType::P256),
             KeyType::C_P384 => Ok(KeyType::P384),
             KeyType::C_P521 => Ok(KeyType::P521),
-            p @ _ => Err(DidPeerError::UnsupportedMulticodecDescriptor(p)),
+            p @ _ => Err(PublicKeyError::UnsupportedMulticodecDescriptor(p)),
         }
     }
 }
