+++ public_key/src/lib.rs
@@ -1,0 +1,7 @@
+mod error;
+mod key;
+mod key_type;
+
+pub use error::PublicKeyError;
+pub use key::Key;
+pub use key_type::KeyType;
